advanced/toots/requirements.txt
Mastodon.py==1.4.2
redis==3.2.1
# hiredis==1.0.0
